Split modular_MLMC tests, moved all fixtures to conftest.py
